объединить компоненты film-details-info и film-details-poster в компоненте film-details-info-wrap
